Highlight search term in results (for catheter). (#37)
If a specific word appears in the search term, then highlight the
results (with Javascript). We would be best doing this highlighting in
the back end, but currently our search returns just the IDs of records
to show, so it would be problematic to get the hightlighted text from
the search results on the page.

Currently, this is triggered by a search containing the word 'catheter',
so for example 'reusable catheter' should highlight the words is matched
(catheter and reusable).